use float32

+++ MachineLearning/Dataset.py
@@ -31,10 +31,10 @@ class CosmoDataset(Dataset):
 
         fn = os.path.join(self.imgdir,
                                 self.frame.iloc[idx, 1])
-        image = io.imread(fn)[np.newaxis,:,:].astype(np.double) / 255
+        image = io.imread(fn)[np.newaxis,:,:].astype(np.float32) / 255
         image = torch.from_numpy( image )
         targets = self.frame.iloc[idx, 4:]
-        targets = np.array([targets]).astype(np.double)
+        targets = np.array([targets]).astype(np.float32)
         targets = torch.from_numpy( targets )
         print( targets )
